Centralize client process watching in the RoslynLanguageServer
