+++ src/LanguageServer/Protocol/Handler/ServerLifetime/InitializeHandler.cs
@@ -2,18 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Diagnostics;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Internal.Log;
+using Microsoft.CommonLanguageServerProtocol.Framework;
 using Roslyn.LanguageServer.Protocol;
 
 namespace Microsoft.CodeAnalysis.LanguageServer.Handler;
 
 [Method(Methods.InitializeName)]
-internal sealed class InitializeHandler() : ILspServiceRequestHandler<InitializeParams, InitializeResult>
+internal sealed class InitializeHandler(ILspLogger logger) : ILspServiceRequestHandler<InitializeParams, InitializeResult>
 {
     public bool MutatesSolutionState => true;
     public bool RequiresLSPSolution => false;
@@ -26,10 +26,11 @@ internal sealed class InitializeHandler() : ILspServiceRequestHandler<Initialize
 
         // If we are running in process with the client, then we don't need to listen for it to exit.
         using var currentProcess = Process.GetCurrentProcess();
-        if (request.ProcessId.HasValue && request.ProcessId.Value != currentProcess.Id)
+        if (request.ProcessId is int clientProcessId
+            && clientProcessId != currentProcess.Id
+            && RoslynLanguageServer.TryRegisterClientProcessId(clientProcessId))
         {
-            // We were given a client process ID. Monitor that process and exit if it exits.
-            _ = WaitForClientProcessExitAsync(request.ProcessId.Value);
+            logger.LogInformation("Monitoring client process {clientProcessId} for exit", clientProcessId);
         }
 
         var capabilitiesProvider = context.GetRequiredLspService<ICapabilitiesProvider>();
@@ -47,29 +48,5 @@ internal sealed class InitializeHandler() : ILspServiceRequestHandler<Initialize
         {
             Capabilities = serverCapabilities,
         };
-    }
-
-    static async Task WaitForClientProcessExitAsync(int clientProcessId)
-    {
-        try
-        {
-            var clientProcessExitTask = new TaskCompletionSource<bool>();
-
-            using var clientProcess = Process.GetProcessById(clientProcessId);
-            clientProcess.EnableRaisingEvents = true;
-            clientProcess.Exited += (sender, args) => clientProcessExitTask.SetResult(true);
-
-            if (!clientProcess.HasExited)
-            {
-                // Wait for the client process to exit.
-                await clientProcessExitTask.Task.ConfigureAwait(false);
-            }
-        }
-        finally
-        {
-            // The process didn't exist, exited, or we ran into
-            // issues checking whether the process had exited.
-            Process.GetCurrentProcess().Kill();
-        }
     }
 }
